# doc/api.yaml
openapi: 3.0.0
info:
  title: WASATEXT
  description: API for WASATEXT messaging platform
  version: 1.0.0
...
security:
  - WASATextAuth: []
servers:
  - url: http://0.0.0.0:3000
    description: Local Server
paths:
...
  /users/{userId}/conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getMessages
      tags:
        - "Conversations"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                description: List of messages
                properties:
                  messages:
                    description: Message schema
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                        $ref: "#/components/schemas/Message"

              example:
                messages:
                  - body:
                      text: "Hello!"
                    messageId: "1"
                    time: "2025-10-31T15:42:00Z"
                    read: "received"
                    sender:
                      userId: "11300"
                      username: Simone

        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Send new direct message
      description: Send new message in the conversation.
      operationId: sendMessage
      tags:
        - "Message"

      parameters:
        - in: query
          name: replyTo
          schema:
            $ref: "#/components/schemas/MessageId"
          description: MessageId of the message we want to reply
          required: false

      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Only text
              properties:
                bodyText:
                  $ref: "#/components/schemas/Text"
              required:
                - bodyText


          multipart/form-data:
            schema:
              type: object
              description: text and/or photo
              properties:
                bodyText:
                  $ref: "#/components/schemas/Text"
                photo:
                  $ref: "#/components/schemas/File"
        required: true
      responses:
        "201":
          description: Message sent. By default, the message has "read" set to sent
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Message"

              example:
                body:
                  text: "Hello!"
                messageId: "4444"
                time: "2025-10-31T15:42:00Z"
                read: "received"

        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /users/{userId}/groups/{groupId}/members/me:
    delete:
      summary: Leave the group
      description: Delete the user from the group.
      operationId: leaveGroup
      tags:
        - "Group"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: OK. Group Left
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups/{groupId}:
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroup
      tags:
        - "Conversations"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: "288344"
                name: "capodanno2026"
                upload:
                  url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                  mime: "image/jpg"
                  width: 1024
                  height: 720
                participants:
                  - userId: "10012"
                    username: "Simone"
                  - userId: "77772"
                    username: "Alessio"


        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/{userId}/groups/{groupId}/messages:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/GroupId"

    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroupMessages
      tags:
        - "Conversations"

      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                description: Response schema. returns che chat
                properties:
                  messages:
                    description: List of messages
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                        $ref: "#/components/schemas/Message"
              example:
                messages:
                  - body:
                      text: "Benvenuti nel gruppo!"
                    messageId: "123456"
                    time: "2025-10-31T15:42:00Z"
                    read: "received"
                    sender:
                      userId: "11300"
                      username: "Simone"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Send new group message
      description: Send new message in the group.
      operationId: sendGroupMessage
      tags:
        - "Message"
      parameters:
        - in : query
          name: replyTo
          schema:
            $ref: "#/components/schemas/MessageId"
          description: MessageId of the message we want to reply
          required: false

      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Only text
              properties:
                bodyText:
                  $ref: "#/components/schemas/Text"
              required:
                - bodyText


          multipart/form-data:
            schema:
              type: object
              description: text and/or photo
              properties:
                bodyText:
                  $ref: "#/components/schemas/Text"
                photo:
                  $ref: "#/components/schemas/File"

        required: true

      responses:
        "201":
          description: Message sent. By default, the message has "read" set to sent.The message has "read" when all members of group has read the message.
          content:
            application/json:
              schema:
                description: Response schema
                allOf:
                  - $ref: "#/components/schemas/Message"

              example:
                body:
                  text : "Hellooooo"
                messageId: "4333744"
                time : 2025-10-31T15:42:00Z
                read: received


        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  schemas:
...
    UserId:
      description: Operation User's ID
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "11300"
      readOnly: true
    Username:
      description: User's username
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[a-zA-Z0-9_ -]+$"
      example: Simone
    User:
      description: User schema
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userId
        - username
    Group:
      type: object
      description: group schema
      properties:
        groupId:
          $ref: "#/components/schemas/GroupId"
        name:
          description: Group name
          type: string
          minLength: 0
          maxLength: 255
          pattern: "^[a-zA-Z0-9_ -]+$"
          example: capodanno2026
        upload:
          $ref: "#/components/schemas/Upload"
        participants:
          description: list of participants of a group
          type: array
          minItems: 2
          maxItems: 256
          items:
            $ref: "#/components/schemas/User"
          example:
            - userId: "11300"
              username: Simone
            - userId: "11301"
              username: Luca
    ConversationId:
      description: Unique ID assigned to each conversation resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "110056"
      readOnly: true
      type: string
...
    MessageId:
      description: Unique ID assigned to each message resource in the system.
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2635444"
      readOnly: true
      type: string
    GroupId:
      description: Unique ID assigned to each group resource in the system.
      type: string
      pattern: "^[0-9]+$"
      minLength: 1
      maxLength: 9999
      example: "2773354"
      readOnly: true
...
    File:
      description: Binary image file
      type: string
      format: binary
      pattern: "^[A-Za-z0-9+/=]+$"
      minLength: 1
      maxLength: 9999999999999999999
    Upload:
      description: upload's schema for responses or view
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-zA-Z]+://.+$"
          format: url
          description: Photo's URL to access the image
          example: https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg
        mime:
          description: mime represents the type of the Photo(image/png...)
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^[a-zA-Z]+/[a-zA-Z.]+$'
          example: image/jpg
        width:
          description: width of the image
          type: integer
          example: 1024
        height:
          description: height of the image
          type: integer
          example: 720
    Message:
      type: object
      description: message schema
      properties:
        messageId:
          $ref: "#/components/schemas/MessageId"
        body:
          type: object
          description: messageContent
          properties:
            text:
              $ref: "#/components/schemas/Text"
            upload:
              $ref: "#/components/schemas/Upload"
        time:
          description: time
          type: string
          format: date-time
          example: "2025-10-31T15:42:00Z"
        sender:
          $ref: "#/components/schemas/User"
        read:
          description: show if the message has been read
          type: string
          enum: [ received, read ]
          example: received
        isForwarded:
          description: show if the message has been forwarded
          type: boolean
          example: false
        replyTo:
          $ref: "#/components/schemas/MessageId"
    Text:
      description: Text of the message
      type: string
      minLength: 0
      maxLength: 2048
      example: Hello, World!
      pattern: "^[a-zA-Z0-9 .,!?_-]+$"
...
    Error:
      description: Info of an error
      type: object
      properties:
        code:
          description: code of the error
          type: integer
          example: 401
        message:
          description: text of the error
          type: string
          minLength: 0
          maxLength: 2048
          example: Unauthorized Action, request invalid or missing.
      required:
        - code
        - message
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User Identifier
      schema:
        $ref: "#/components/schemas/UserId"
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Chat Identifier
      schema:
        $ref: "#/components/schemas/ConversationId"
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group Identifier
      schema:
        $ref: "#/components/schemas/GroupId"
  securitySchemes:
    WASATextAuth:
      type: http
      description: Protect operation session token.
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    UnauthorizedError:
      description: Unauthorized Action, request invalid or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: "Unauthorized Action, request invalid or missing."
    InternalError:
      description: An error occurred, please retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: "An error occurred, please retry."
    NotFoundError:
      description: Resource not found. Check the parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: "Resource not found. Check the parameters."
    BadRequest:
      description: Bad Request. Invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 400
            message: "Bad Request. Invalid parameters."
...